pyproject.toml
[tool.poetry]
name = "finformer"
version = "0.1.0"
description = ""
authors = ["Hleb Stenin <hlebstenin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
tqdm = "^4.66.2"
yfinance = "^0.2.37"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
